final_requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
braceexpand==0.1.7
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.0
ctranslate2==4.1.0
cycler==0.12.1
Cython==3.0.9
cytoolz==0.12.3
datasets
decorator==5.1.1
deepmultilingualpunctuation==1.0.1
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
dora_search==0.1.12
editdistance==0.8.1
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
faster-whisper==1.0.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec
g2p-en==2.1.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
grpcio==1.62.1
huggingface-hub
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.6
inflect==7.0.0
intervaltree==3.1.0
ipython==8.22.2
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.3
joblib==1.3.2
julius==0.2.7
jupyterlab_widgets==3.0.10
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
lameenc==1.7.0
lazy_loader==0.3
Levenshtein==0.25.0
lhotse==1.22.0
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.10.1
lilcom==1.7
llvmlite==0.42.0
loguru==0.7.2
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nemo-toolkit==1.21.0
networkx==3.2.1
nltk==3.8.1
noisereduce==3.0.2
numba==0.59.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.17.1
openunmix==1.2.1
optuna==3.5.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
plac==1.4.3
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==15.0.1
pyarrow-hotfix==0.6
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.14
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pysrt==1.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.7
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
sentry-sdk==1.42.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
sox==1.4.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.28
stack-data==0.6.3
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
text-unidecode==1.3
texterrors==0.4.4
threadpoolctl==3.3.0
tokenizers==0.15.2
toolz==0.12.1
torch
torch-audiomentations
torch-pitch-shift
torchaudio
torchmetrics
torchvision
tqdm==4.66.2
traitlets==5.14.2
transformers
treetable==0.2.5
triton
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.4
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==3.0.1
wget==3.2
-e git+https://github.com/m-bain/whisperX@78dcfaab51005aa703ee21375f81ed31bc248560#egg=whisperx
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtokentome==1.0.6
numpy>=1.22,<1.24
